refactor: reorganized queries
