+++ src/collection/queries/pipeline/pipeline-stage.ts
